Create classes for the providers
