+++ .pylintrc
@@ -26,7 +26,8 @@ disable=
     too-many-statements,
     wrong-import-order,
     too-many-branches,
-    import-error
+    import-error,
+    duplicate-code
 
 
 [REPORTS]
