--- mtgjson4/provider/__init__.py
@@ -1,1 +1,0 @@
-"""Upstream data providers."""
