+++ mtgjson4/providers/__init__.py
@@ -1,0 +1,12 @@
+"""Upstream data providers."""
+from mtgjson4.providers.gamepedia import Gamepedia
+from mtgjson4.providers.gatherer import Gatherer
+from mtgjson4.providers.scryfall import Scryfall
+from mtgjson4.providers.tcgplayer import TCGPlayer
+from mtgjson4.providers.wizards import Wizards
+
+GATHERER = Gatherer()
+SCRYFALL = Scryfall()
+TCGPLAYER = TCGPlayer()
+WIZARDS = Wizards()
+GAMEPEDIA = Gamepedia()
